requirements.txt
streamlit==1.37.1
supabase==2.5.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pandas==2.2.2
shortuuid==1.0.13
httpx==0.27.0
pydantic==2.7.4
pyngrok
